perf: Move `model_dump` to `BaseModel`
This allows us to delete the dict comprehension from
`scim_serializer` and pydantic's none exclusion preserved
